// tv-projector/src/types/api.ts
export type LiveStream = {
  id: string;
  title: string;
  thumbnailUrl: string | null;
  viewerCount: number;
  category: string;
  streamer: string;
  streamerSlug: string;
  avatarUrl: string | null;
  isPaid?: boolean;
  entryCoinCost?: number | null;
  entryPriceUsd?: number | null;
};